feat: Replace broken logo image with inline SVG
The previous logo image was broken and not loading. This commit replaces it with an inline SVG representation of the Juju Kids logo. The SVG is designed to be white, simplifying styling and ensuring it displays correctly. CSS filters that were previously used to achieve a white logo from a potentially colored image have been removed as they are no longer necessary.